Separated SimulationRunner class from Simulator class for code clarity.
